+++ lib/pavlov/alpha_compatibility.rb
@@ -1,4 +1,9 @@
+require 'active_support/concern'
+require 'active_support/inflector'
 require 'pavlov'
+
+# Everything in this file should be considered deprecated. It will go away
+# sometime before 1.0
 
 module Pavlov
   def self.old_command command_name, *args
@@ -62,7 +67,41 @@ module Pavlov
     end
   end
 
+  module Helpers
+    def interactor name, *args
+      args = add_pavlov_options args
+      Pavlov.interactor name, *args
+    end
+
+    def query name, *args
+      args = add_pavlov_options args
+      Pavlov.query name, *args
+    end
+
+    def command name, *args
+      args = add_pavlov_options args
+      Pavlov.command name, *args
+    end
+
+    def pavlov_options
+      {}
+    end
+
+    private
+    def add_pavlov_options args
+      # TODO: we should do this at a point where we know how many arguments we need
+      # so we can decide if we need to merge with another options object or
+      # just add it.
+      if pavlov_options != {}
+        args << pavlov_options
+      end
+      args
+    end
+  end
+
   module Operation
+    include Pavlov::Helpers
+
     module ClassMethods
       def arguments(*args)
         # Add generic attribute for each argument
@@ -75,4 +114,35 @@ module Pavlov
       end
     end
   end
+
+  module Command
+    extend ActiveSupport::Concern
+    include Pavlov::Operation
+  end
+
+  module Query
+    extend ActiveSupport::Concern
+    include Pavlov::Operation
+  end
+
+  module Interactor
+    extend ActiveSupport::Concern
+    include Pavlov::Operation
+
+    module ClassMethods
+      # make our interactors behave as Resque jobs
+      def perform(*args)
+        new(*args).call
+      end
+
+      def queue
+        @queue ||= :interactor_operations
+      end
+    end
+
+    def authorized?
+      raise NotImplementedError
+    end
+  end
+
 end