+++ CHANGELOG.md
@@ -12,8 +12,13 @@ This release brings forth lots and lots of incompatibilities. Where possible, we
 
 If you want to retain deprecated functionality, you can `require 'pavlov/alpha_compatibility'`.
 
+* Deprecated `Pavlov.query` etc.
+* Deprecated `Pavlov::Helpers`.
 * Deprecated `arguments` in operations.
 * Deprecated `pavlov_options` that were used by the helpers.
+* Deprecated `Pavlov::Interactor` in favor of the pre-generated `Interactor` class
+* Deprecated `Pavlov::Command` in favor of the pre-generated `Command` class
+* Deprecated `Pavlov::Query` in favor of the pre-generated `Query` class
 
 #### Completely removed: